# backend/src/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Medilink Healthcare API",
    "description": "Comprehensive Interactive API Documentation for Medilink Healthcare Platform",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local Development Server"
    },
    {
      "url": "/",
      "description": "Relative server route"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase JWT authentication token."
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/api/appointments/create": {
      "post": {
        "summary": "Book Appointment",
        "tags": ["Appointments"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["full_name", "age", "phone", "department", "preferred_date", "preferred_time"],
                "properties": {
                  "full_name": { "type": "string", "example": "Manohar Sameer" },
                  "age": { "type": "integer", "example": 35 },
                  "phone": { "type": "string", "example": "9123456780" },
                  "department": { "type": "string", "example": "Cardiology" },
                  "preferred_date": { "type": "string", "example": "2026-08-15" },
                  "preferred_time": { "type": "string", "example": "10:30" },
                  "symptoms": { "type": "string", "example": "Mild chest tightness" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Appointment registered" }
        }
      }
    },
    "/api/appointments/track": {
      "post": {
        "summary": "Track Appointment Reference",
        "tags": ["Appointments"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "search": { "type": "string", "example": "APT-1786333180" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Success" }
        }
      }
    },
...
  }
}